xorg: Mark obsolete xorgvideo* drivers broken
These are now showing up as broken builds in Hydra since 2daefaf4572.
None of these compiled even in 16.03 and I think all of them are
for pretty obsolete hardware, so just mark them as broken.

(In principle the xorg generator could be made to ignore them but that
would be more work.)