linux_rpi: More downstream device tree name fixups
Also remove the upstream device trees beforehand so we don't
accidentally end using them if we mess up in postFixup.

Also on the topic of mess ups, make the copy unconditional since they
always seem to be built on all architectures anyway so checking for
their existence just hides typos.