stdenv: Make it easier to correctly run phases in nix-shell
Currently, it's a bit annoying to run build phases manually inside a
nix-shell. To run say, buildPhase, the invocation that works correctly
in all cases is:

eval "${buildPhase:-buildPhase}"

Which is quite a mouthful and non-obvious. Even worse, the obvious
thing to try, i.e.:

buildPhase

works _sometimes_ (i.e. when the package doesn't have a
`buildPhase = "...";`) but silently does the wrong thing in other cases.

Let's fix this by having the obvious thing work always.

To do that, we simply pull the logic of 'if $buildPhase exists as a variable,
then eval "$buildPhase", else do the default buildPhase' into the
buildPhase() function itself using a common helper:

buildPhase() {
    commonPhaseImpl buildPhase --default defaultBuildPhase --pre-hook preBuild --post-hook postBuild
}

As a bonus, we get to fix another current annoyance/footgun at the same
time, that is setting e.g. `buildPhase = "...";` makes the preBuild and
postBuild hooks not being run anymore.